requirements.txt
redis
python-dotenv
pika